Conflicts:
	lib/files/storage/mappedlocal.php

+++ apps/files_external/lib/amazons3.php
@@ -33,10 +33,25 @@ use Aws\S3\Exception\S3Exception;
 
 class AmazonS3 extends \OC\Files\Storage\Common {
 
+	/**
+	 * @var \Aws\S3\S3Client
+	 */
 	private $connection;
+	/**
+	 * @var string
+	 */
 	private $bucket;
+	/**
+	 * @var array
+	 */
 	private static $tmpFiles = array();
+	/**
+	 * @var bool
+	 */
 	private $test = false;
+	/**
+	 * @var int
+	 */
 	private $timeout = 15;
 
 	private function normalizePath($path) {
@@ -64,7 +79,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 
 		$this->bucket = $params['bucket'];
 		$scheme = ($params['use_ssl'] === 'false') ? 'http' : 'https';
-		$this->test = ( isset($params['test'])) ? true : false;
+		$this->test = isset($params['test']);
 		$this->timeout = ( ! isset($params['timeout'])) ? 15 : $params['timeout'];
 		$params['region'] = ( ! isset($params['region'])) ? 'eu-west-1' : $params['region'];
 		$params['hostname'] = ( !isset($params['hostname'])) ? 's3.amazonaws.com' : $params['hostname'];
@@ -96,6 +111,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 				));
 			$this->testTimeout();
 			} catch (S3Exception $e) {
+				\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 				throw new \Exception("Creation of bucket failed.");
 			}
 		}
@@ -129,6 +145,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 			));
 			$this->testTimeout();
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 
@@ -150,6 +167,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 				$path
 			);
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 
@@ -165,17 +183,20 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 		}
 
 		$dh = $this->opendir($path);
-		while ($file = readdir($dh)) {
-			if ($file === '.' || $file === '..') {
-				continue;
-			}
 
-			if ($this->is_dir($path . '/' . $file)) {
-				$this->rmdir($path . '/' . $file);
-			} else {
-				$this->unlink($path . '/' . $file);
+		if(is_resource($dh)) {
+			while (($file = readdir($dh)) !== false) {
+				if ($file === '.' || $file === '..') {
+					continue;
+				}
+
+				if ($this->is_dir($path . '/' . $file)) {
+					$this->rmdir($path . '/' . $file);
+				} else {
+					$this->unlink($path . '/' . $file);
+				}
 			}
-               	}
+		}
 
 		try {
 			$result = $this->connection->deleteObject(array(
@@ -184,6 +205,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 			));
 			$this->testTimeout();
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 
@@ -221,6 +243,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 
 			return opendir('fakedir://amazons3' . $path);
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 	}
@@ -249,6 +272,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 
 			return $stat;
 		} catch(S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 	}
@@ -269,6 +293,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 				return 'dir';
 			}
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 
@@ -293,6 +318,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 			));
 			$this->testTimeout();
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 
@@ -315,6 +341,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 						'SaveAs' => $tmpFile
 					));
 				} catch (S3Exception $e) {
+					\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 					return false;
 				}
 
@@ -361,6 +388,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 					'Key' => $path
 				));
 			} catch (S3Exception $e) {
+				\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 				return false;
 			}
 
@@ -398,6 +426,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 				$this->testTimeout();
 			}
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 
@@ -417,6 +446,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 				));
 				$this->testTimeout();
 			} catch (S3Exception $e) {
+				\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 				return false;
 			}
 		} else {
@@ -432,19 +462,22 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 				));
 				$this->testTimeout();
 			} catch (S3Exception $e) {
+				\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 				return false;
 			}
 
 			$dh = $this->opendir($path1);
-			while ($file = readdir($dh)) {
-				if ($file === '.' || $file === '..') {
-					continue;
-				}
+			if(is_resource($dh)) {
+				while (($file = readdir($dh)) !== false) {
+					if ($file === '.' || $file === '..') {
+						continue;
+					}
 
-				$source = $path1 . '/' . $file;
-				$target = $path2 . '/' . $file;
-				$this->copy($source, $target);
-                	}
+					$source = $path1 . '/' . $file;
+					$target = $path2 . '/' . $file;
+					$this->copy($source, $target);
+				}
+			}
 		}
 
 		return true;
@@ -514,6 +547,7 @@ class AmazonS3 extends \OC\Files\Storage\Common {
 
 			unlink($tmpFile);
 		} catch (S3Exception $e) {
+			\OCP\Util::writeLog('files_external', $e->getMessage(), \OCP\Util::ERROR);
 			return false;
 		}
 	}
